code-monitor/requirements.txt
GitPython==3.1.43
requests==2.31.0
python-dotenv==1.0.1
dataclasses-json==0.6.4